Exclude "sharp" dependency from deployment bundle
Next.js v10 adds "sharp" (an image processing module) as a dependency. It is > 50 MB in size uncompressed, and was bloating our deployment bundle from ~15 MB to ~38 MB (compressed). The deployment bundle size negatively affects the Lambda's (and therefore Course Search's) performance. Sharp isn't needed at runtime (I'm fairly sure) so it can be excluded from the deployment bundle. I have deployed this version to my sandbox and verified that the app behaves as it should.